+++ Calculator/Calculator/Formula.swift
@@ -2,10 +2,18 @@ import Foundation
 
 struct Formula {
     var operands = CalculatorItemQueue<Double>()
-    var operators = CalculatorItemQueue<String>()
+    var operators = CalculatorItemQueue<Character>()
     
     func result() -> Double {
-
+        var result: Double = 0
+        while operands.linkedList.head == nil {
+            let value = operands.deleteFromQueue()
+            guard let oper = Operator(rawValue: operators.deleteFromQueue()) else {
+                return 0      // 에러처리 구문 추가 시 수정
+            }
+            result = oper.calculate(lhs: result, rhs: value)
+        }
+        return result
     }
 }
 
@@ -17,32 +25,5 @@ extension String {
 }
 
 
-enum ExpressionParser {
-    func parse(from input: String) -> Formula {
-        let operatorEnumArray = Operator.allCases.map { String($0.rawValue) }
-        let inputCharacters = Array(input).map { String($0) }
-        let inputOperator = inputCharacters.filter { operatorEnumArray.contains($0) == true }
-        let inputOperands = componentsByOperators(from: input)
-        
-        let formula = Formula()
-        for operandIndex in 0...inputOperands.count {
-            formula.operands.insertToQueue(Double(inputOperands[operandIndex]) ?? 0)
-        }
-        for operatorIndex in 0...inputOperator.count {
-            formula.operators.insertToQueue(inputOperator[operatorIndex])
-        }
-        return formula
-    }
 
-    private func componentsByOperators(from input: String) -> [String] {
-        var splitedWithTarget: [String]
-        var joinedString = input
-        for operatorCase in Operator.allCases {
-            splitedWithTarget = joinedString.split(with: operatorCase.rawValue)
-            joinedString = splitedWithTarget.joined(separator: " ")
-        }
-        let inputOperands = joinedString.components(separatedBy: " ")
-        return inputOperands
-    }
-}
